### 2019-02-20 #1268 @whedon
[PRE REVIEW]: uJVM: Lightweight Java Virtual Machine for embedded systems
**Submitting author:** @omoliavko (<a href="http://orcid.org/0000-0003-3167-983X">Oleksandr Moliavko</a>)
**Repository:** <a href="https://github.com/Samsung/uJVM/" target ="_blank">https://github.com/Samsung/uJVM/</a>
**Version:** v0.1
