### 2019-02-20 #1268 @arfon
:wave: @omoliavko - please review the submission guidelines on how to prepare your JOSS paper. In particular, your paper doesn't seem to have an associated bibtex file: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
