### 2019-02-20 #1268 @omoliavko
Hello, as of now we don't have any publications regarding uJVM; therefore, we can't provide any bibtex file. Is it strictly necessary for publication? If yes, can we provide empty bibtex file?
