### 2019-02-20 #1268 @arfon
You don't need to cite publications, you should move what you currently have in this section into the bibtex file and cite accordingly: https://github.com/Samsung/uJVM/blob/master/docs/joss/paper.md#references

Also, any related work should be cited too.
