### 2019-02-20 #1268 @omoliavko
Thank you for quick answer; we'll prepare related works list and add bibtex file with citations ASAP
