### 2019-02-21 #1268 @omoliavko
Hello, I've created a separate paper.bib file with references to materials we've used during uJVM development and updated paper.md file accordingly.
