### 2019-02-21 #1268 @arfon
@omoliavko - please merge this PR: https://github.com/Samsung/uJVM/pull/63

Also, you need to cite the entries in your bibtex directly please. (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
