### 2019-02-21 #1268 @omoliavko
Merge done; also, I've updated reference marks in paper.md
