### 2019-02-21 #1268 @arfon
Great, thanks for your work getting the paper to compile. Reading this paper, it's not immediately clear to me what the research challenge is that you're addressing here.

Quoting from our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements):

> - The software should have an obvious research application
> - You should be a major contributor to the software you are submitting
> - The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)
> - The software should be feature complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications). Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable.

Could you say a little more in your paper about what sort of research challenges this software is designed to address.

Also, I'm cc-ing @gkthiruvathukal who I think would be a good editor for this submission.
