### 2019-02-21 #1268 @omoliavko
Many thanks for your patience in explaining how to make paper valid. I've updated the text to explain potential research application in a bit more detail.
