### 2019-03-12 #1268 @omoliavko
Is there something I need to do for paper review to actually start or I should just wait for assigned reviewer?
