### 2019-03-13 #1268 @gkthiruvathukal
@omoliavko  In looking more closely at the review criteria at https://joss.readthedocs.io/en/latest/review_criteria.html, I was wondering whether you either comment on or reviews your paper to include how your contribution addresses "any ongoing research projects using the software or recent scholarly publications enabled by it."
