### 2019-03-13 #1268 @omoliavko
@gkthiruvathukal I've rewritten the part of article regarding potential research applications of uJVM. Could you please take another look at it?
