### 2019-03-14 #1268 @omoliavko
Based on data from the list, I think the following could be reviewers for paper: hainesr, jizecn, adavidzh, dvysardana.
I've updated the paper text by extending list of supported H/W and RTOSes, but it should not have significant effect on research-related aspects.
