### 2019-03-14 #1268 @gkthiruvathukal
@hainesr Would you be willing to contribute a review for JOSS on this submission? This submission seems closely related to your areas of expertise and interest.
