### 2019-03-15 #1268 @gkthiruvathukal
@jizecn and @adavidzh: Would either or both of you be willing to contribute a review for JOSS on this submission? I haven't heard back from @hainesr yet. Please let me know so we can begin review of this submission.
