### 2019-03-19 #1268 @gkthiruvathukal
@jmkim and @morganericsson, would either or both of you be willing to contribute a review to this JOSS submission? As you can see in this issue thread, I have asked others who are either not available or have not responded. Both of you have Java listed and systems/software engineering interests, so I think this submission would be of interest to you.
