### 2019-03-20 #1338 @hainesr
Hi @omoliavko, should `vpetrychenko` also be an author on this paper? They appear to have made a significant contribution to the software - https://github.com/Samsung/uJVM/graphs/contributors
