### 2019-03-20 #1338 @omoliavko
Hi @hainesr, I could add him to authors list, but he contributed to S/W, not the paper itself, so I'm not sure if it would be correct.
