### 2019-03-20 #1338 @hainesr
The software is the important contribution here, so I think he should be added. The paper is of secondary importance where authorship is concerned. I'll tag in @gkthiruvathukal for an adjudication, though...
