### 2019-03-20 #1338 @gkthiruvathukal
I concur with @hainesr. I would think that one of the main goals of JOSS is to ensure that those who contribute substantially to the development of research software are recognized for their contributions.
