### 2019-03-20 #1338 @omoliavko
Added Vitalii Petrychenko (vpetrychenko) to author list
