### 2019-03-24 #1338 @hainesr
Hi @omoliavko, I had some issues building this project with my setup. I have JDK 11 as my default java, but setting `$JAVA_HOME` to point to JDK 8 didn't have any effect. I have submitted a pull request (Samsung/uJVM#79) to fix this.
