### 2019-03-25 #1338 @tdrozdovsky
Great job! I've merged your pull request. We will add a description "how to set JAVA_HOME for Java "
