### 2019-04-01 #1338 @morganericsson
@arfon I've had some issues with the software. Since this is my first review for JOSS, I am unsure where/how to voice my concerns, but I'll do it here and then we can edit if it is the wrong place.

- The current version seems to be 0.2, so my comments are based on that version.
- It is not clear who this is intended for, see the next item...
- The documentation (apart from compiling) is quite poor, so it is difficult to judge exactly what the current release can do. I tried to write a few simple programs, and quickly ran into multiple difficulties, e.g.:
   - It was not obvious how to package and run programs.
   - It was not obvious which classes/functions I could use. Again, the japps and tests helped, but the documentation really needs to be better.
  - I ran into multiple failures (e.g., `Ret <something> @ instr right before <addr>`) from even simple loops and methods.

Based on this, I am not sure if the submission fulfills "The software should be feature complete" (and "either enables some new research challenges to be addressed or makes addressing research challenges significantly better" but I guess this depends on the target audience).
