### 2019-04-02 #1338 @omoliavko
@morganericsson Thanks for the valuable feedback, we'll improve the documentation as soon as feasible. Now answers to your specific comments:
1) Current version is 0.2 - we're actively developing the uJVM, taking feedback from multiple sources into account; indeed, you're welcome to give feedback if you have time and feel inclined to do so
2) We'll update the article to specify intended applications of uJVM; however, right now it is more of an experimental development than application meant for release
3.1) We'll provide documentation that describes the process of creating applications and running them on uJVM;
3.2) As of March 2019, uJVM development is centered on creating efficient Java interpreter, to prove the feasibility of Java apps for embedded systems - indeed, this was the primary objective of uJVM research project from the very beginning; creating extended library of classes similar to standard Java library is only in plans; we'll provide list of currently supported library classes and methods in documentation
3.3) We'd be grateful if you could send detailed error reports to us or create issues on github so we could analyze and fix the bugs you've encountered
Once again, thanks for your time and effort
