### 2019-04-02 #1338 @omoliavko
@hainesr @morganericsson Javadoc documentation generation is added to uJVM to view info on supported Java classes; updated README.md to reflect this and added uJVM_functionality_description.md file to docs folder.
