### 2019-04-03 #1338 @omoliavko
@hainesr  @morganericsson A guide on building apps for uJVM was added - please see docs/java_app_creating_guide.md file. We'd be grateful for your opinion on it.
