### 2019-04-05 #1338 @morganericsson
I've reviewed the added files and the updated paper, and updated my review checklist accordingly. Overall, I think it's fine and all of my concerns have been addressed. The documentation, of course, should improve over time, but I think it's sufficient for now.
