### 2019-04-10 #1338 @hainesr
I have raised a new PR to fix javadoc on setups with multiple JDKs: Samsung/uJVM#85
