### 2019-04-10 #1338 @omoliavko
@hainesr Pull Request accepted, thank you for collaboration
