### 2019-04-10 #1338 @hainesr
My checklist is complete now @gkthiruvathukal, @arfon. I'm happy with the submission!

I think a new release now needs to be tagged @omoliavko. Would you consider calling it 1.0 in line with Semantic Versioning? https://semver.org/

One minor nit-pick: in the paper itself, I think it would read better if you wrote 'software' and 'hardware' out longhand instead of using 'S/W' and 'H/W'. But this is just my personal preference and not necessarily a barrier to acceptance.
