### 2019-04-10 #1338 @omoliavko
@hainesr We're also happy with our progress so far; unfortunately, we cannot tag a new release now, because we have a development plan to adhere to, and next release of uJVM shall include several features that are in development now.
