### 2019-04-13 #1338 @tdrozdovsky
@arfon @kyleniemeyer @gkthiruvathukal Could you please explain what steps we should take now that @morganericsson and @hainesr have completed the checklist?
