### 2019-04-13 #1338 @kyleniemeyer
At this point @gkthiruvathukal will give it a final look as the editor for the submission; if he decides it is ready to accept, then he'll ask you to archive the accepted version of the package and provide an associated DOI (e.g., via Zenodo). However, hold off until you get that request from him.
