### 2019-04-14 #1338 @gkthiruvathukal
@kyleniemeyer I'm on travel right now (a mix of business and pleasure). I will work on the decision in the next day or two.  I'm encouraged by what I am seeing though. Thanks for your patience. I will try not to keep you waiting for too long.
