### 2019-04-16 #1338 @gkthiruvathukal
@tdrozdovsky and @omoliavko, I have gone through the checklist, and I'm ready to accept. Based on @kyleniemeyer's follow-up, can you please archive the accepted version of the package and provide an associated DOI (e.g., via Zenodo).

@kyleniemeyer Just to clarify, I do not do an accept via Whedon just yet, right?
