### 2019-04-16 #1338 @danielskatz
Hi @gkthiruvathukal - rather than asking a specific AEiC, you probably should ask all of us (to get the person on duty this week) by including @openjournals/joss-eics in the comment.

And in this case, no, don't do an accept yet - you can do that once these other things (version, archive) are done, or the AEiC can do it.
