### 2019-04-16 #1338 @gkthiruvathukal
@danielskatz Thanks for clarifying how to ask all AEICs. Somehow that was lost on me.

@openjournals/joss-eics This submission is moving toward acceptance. As this is one of the first I have seen through the process, it would be good if you can take a look that all is in order. The checklists are fine and all issues seem to have been addressed in the issue thread and follow-up. I believe we just need the DOI for the accepted version of the package.
