### 2019-04-16 #1338 @arfon
> The checklists are fine and all issues seem to have been addressed in the issue thread and follow-up. I believe we just need the DOI for the accepted version of the package.

Yep, I think we're in good shape here. At this point we just need the author to create a DOI for the software and post the URL to it here.
