### 2019-04-19 #1338 @omoliavko
@gkthiruvathukal I've posted uJVM source code archive to Zenodo: https://zenodo.org/record/2647414
