### 2019-04-19 #1338 @gkthiruvathukal
@openjournals/joss-eics I believe we are close to acceptance now. Are the authors supposed to cite the Zenodo artifact in the list of references in the accepted paper?
