### 2019-04-19 #1338 @arfon
> @openjournals/joss-eics Thanks for all the help. I think I have a good handle on the workflow now.

:+1: are you happy with the final proofs @gkthiruvathukal? If so, I can go ahead and process the paper acceptance.
