### 2019-04-19 #1338 @gkthiruvathukal
@arfon The proofs looked fine when I checked. I'm mobile now but am confident we can process acceptance.
