### 2019-09-20 #1338 @danielskatz
This paper has been retracted by the JOSS editorial team due to a [DMCA takedown](https://github.com/github/dmca/blob/881963da39208135e29a50847c3abe9afe683e0a/2019/07/2019-07-29-uJVM.md).
